db_service/requirements.txt
mysql-connector-python
kafka-python==2.0.2
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0